requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.9.0
branca==0.8.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
decorator==5.2.1
exceptiongroup==1.2.2
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
folium==0.19.5
fonttools==4.57.0
fsspec==2025.3.2
future==1.0.0
geocoder==1.38.1
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grad-cam==1.5.5
gradio_client==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
orjson==3.10.16
packaging==25.0
pandas==2.2.3
pillow==11.0.0
proto-plus==1.26.1
protobuf==6.30.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
pynmea2==1.19.0
pyparsing==3.2.3
pyperclip==1.9.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pywin32==310
PyYAML==6.0.2
ratelim==0.1.6
requests==2.32.3
rich==14.0.0
rsa==4.9.1
ruff==0.11.7
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.1
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2024.8.30
tkintermapview==1.29
tomlkit==0.12.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchcam==0.4.0
torchvision==0.20.1+cu121
tqdm==4.67.1
ttach==0.0.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics @ file:///C:/Users/bentl/Desktop/THESIS_ROAD-DEFECT/Road-Defect-Indexing-System/yolov10
ultralytics-thop==2.0.14
urllib3==2.4.0
uvicorn==0.34.2
websockets==12.0
xyzservices==2025.4.0
zipp==3.21.0
